services/bot/requirements.txt
python-telegram-bot>=13.11
yt-dlp>=2022.6.29
loguru~=0.6.0
botocore~=1.27.13
boto3~=1.24.13
requests~=2.28.1
validators~=0.20.0
pytz~=2022.7